nidigital example updates
