+++ examples/nidigital_spi/measurement.py
@@ -3,7 +3,7 @@
 import contextlib
 import logging
 import pathlib
-from typing import Any, Dict, Iterable, Tuple, Union
+from typing import Iterable, Tuple, Union
 
 import click
 import grpc
@@ -11,11 +11,14 @@ import nidigital
 from _helpers import (
     ServiceOptions,
     configure_logging,
+    create_session_management_client,
+    get_grpc_device_channel,
     get_service_options,
     grpc_device_options,
-    verbosity_option,
     use_simulation_option,
+    verbosity_option,
 )
+from _nidigital_helpers import _create_nidigital_session
 
 import ni_measurementlink_service as nims
 
@@ -58,12 +61,7 @@ def measure(
     """Test a SPI device using an NI Digital Pattern instrument."""
     logging.info("Starting test: pin_names=%s", pin_names)
 
-    session_management_client = nims.session_management.Client(
-        grpc_channel=measurement_service.get_channel(
-            provided_interface=nims.session_management.GRPC_SERVICE_INTERFACE_NAME,
-            service_class=nims.session_management.GRPC_SERVICE_CLASS,
-        )
-    )
+    session_management_client = create_session_management_client(measurement_service)
 
     with contextlib.ExitStack() as stack:
         reservation = stack.enter_context(
@@ -85,7 +83,8 @@ def measure(
             )
 
         session_info = reservation.session_info[0]
-        session = stack.enter_context(_create_nidigital_session(session_info))
+        grpc_device_channel = get_grpc_device_channel(measurement_service, nidigital)
+        session = stack.enter_context(_create_nidigital_session(session_info, grpc_device_channel))
 
         pin_map_context = measurement_service.context.pin_map_context
         selected_sites_string = ",".join(f"site{i}" for i in pin_map_context.sites)
@@ -116,36 +115,6 @@ def measure(
 
     logging.info("Completed test: passing_sites=%s failing_sites=%s", passing_sites, failing_sites)
     return (passing_sites, failing_sites)
-
-
-def _create_nidigital_session(
-    session_info: nims.session_management.SessionInformation,
-) -> nidigital.Session:
-    options: Dict[str, Any] = {}
-    if service_options.use_simulation:
-        options["simulate"] = True
-        options["driver_setup"] = {"Model": "6570"}
-
-    session_kwargs: Dict[str, Any] = {}
-    if service_options.use_grpc_device:
-        session_grpc_address = service_options.grpc_device_address
-
-        if not session_grpc_address:
-            session_grpc_channel = measurement_service.get_channel(
-                provided_interface=nidigital.GRPC_SERVICE_INTERFACE_NAME,
-                service_class="ni.measurementlink.v1.grpcdeviceserver",
-            )
-        else:
-            session_grpc_channel = measurement_service.channel_pool.get_channel(
-                target=session_grpc_address
-            )
-        session_kwargs["grpc_options"] = nidigital.GrpcSessionOptions(
-            session_grpc_channel,
-            session_name=session_info.session_name,
-            initialization_behavior=nidigital.SessionInitializationBehavior.AUTO,
-        )
-
-    return nidigital.Session(session_info.resource_name, options=options, **session_kwargs)
 
 
 def _resolve_relative_path(
